Raise when primary keys in Ecto are nils
In some cases, when we have an Ecto payload, primary key can be set to nil. This can lead to situation where dataloader is associating the incorrect child object to a parent object. (See the issue here: https://github.com/absinthe-graphql/dataloader/issues/172).

This helps to avoid overlooking such situations by raising when nil was found in primary keys.